Use `int32_t` type for most transaction size/weight values
This change gets rid of a few casts and makes the following commit diff
smaller.

+++ src/txmempool.h
@@ -222,7 +222,7 @@ struct TxMempoolInfo
     CAmount fee;
 
     /** Virtual size of the transaction. */
-    size_t vsize;
+    int32_t vsize;
 
     /** The fee delta. */
     int64_t nFeeDelta;
